+++ orcid-core/src/main/java/org/orcid/core/messaging/JmsMessageSender.java
@@ -4,8 +4,8 @@ import java.util.Map;
 
 import javax.annotation.Resource;
 
-import org.orcid.utils.listener.LastModifiedMessage;
-import org.orcid.utils.solr.entities.OrgDisambiguatedSolrDocument;
+import org.orcid.core.utils.listener.LastModifiedMessage;
+import org.orcid.core.utils.solr.entities.OrgDisambiguatedSolrDocument;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.jms.JmsException;
