+++ orcid-core/src/main/java/org/orcid/core/utils/listener/RetryMessage.java
@@ -1,6 +1,9 @@
-package org.orcid.utils.listener;
+package org.orcid.core.utils.listener;
 
 import java.util.Map;
+
+import org.orcid.core.utils.listener.BaseMessage;
+import org.orcid.core.utils.listener.MessageConstants;
 
 import com.google.common.collect.ImmutableMap;
