+++ orcid-core/src/main/java/org/orcid/core/utils/solr/entities/OrcidSolrDocument.java
@@ -1,4 +1,4 @@
-package org.orcid.utils.solr.entities;
+package org.orcid.core.utils.solr.entities;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -19,7 +19,7 @@ import org.apache.solr.client.solrj.beans.Field;
  * 
  */
 
-public class OrcidSolrDocumentLegacy {
+public class OrcidSolrDocument {
 
     @Field
     private String orcid;
@@ -36,6 +36,9 @@ public class OrcidSolrDocumentLegacy {
     @Field(SolrConstants.CREDIT_NAME)
     private String creditName;
 
+    @Field(SolrConstants.BIOGRAPHY)
+    private String biography;
+    
     @Field(SolrConstants.OTHER_NAMES)
     private List<String> otherNames;
 
@@ -43,8 +46,14 @@ public class OrcidSolrDocumentLegacy {
     private List<String> externalIdSources;
 
     @Field(SolrConstants.EXTERNAL_ID_REFERENCES)
-    private List<String> externalIdReferences;
+    private List<String> externalIdReferences;    
 
+    @Field(SolrConstants.EXTERNAL_ID_SOURCE_AND_REFERENCE)
+    private List<String> externalIdSourceReferences;
+    
+    @Field(SolrConstants.EXTERNAL_ID_TYPE_AND_VALUE)
+    private List<String> externalIdTypeAndValue;
+    
     @Field(SolrConstants.DIGITAL_OBJECT_IDS)
     private List<String> digitalObjectIds = new ArrayList<String>();;
 
@@ -170,6 +179,15 @@ public class OrcidSolrDocumentLegacy {
 
     @Field(SolrConstants.PRIMARY_RECORD)
     private String primaryRecord;
+    
+    @Field(SolrConstants.PEER_REVIEW_GROUP_ID)
+    private Set<String> peerReviewGroupId;
+    
+    @Field(SolrConstants.PEER_REVIEW_ROLE)
+    private Set<String> peerReviewRole;
+    
+    @Field(SolrConstants.PEER_REVIEW_TYPE)
+    private Set<String> peerReviewType;
 
     // e.g. doi-self
     @Field("*" + SolrConstants.DYNAMIC_SELF)
@@ -196,7 +214,13 @@ public class OrcidSolrDocumentLegacy {
     
     @Field(SolrConstants.RESEARCH_RESOURCE_PROPOSAL_TITLES)
     private List<String> researchResourceProposalTitles;
-
+    
+    @Field(SolrConstants.AFFILIATE_PAST_INSTITUTION_NAMES)
+    private List<String> pastInstitutionAffiliationNames;
+    
+    @Field(SolrConstants.AFFILIATE_CURRENT_INSTITUTION_NAME)
+    private List<String> currentInstitutionAffiliationNames;
+    
     public String getOrcid() {
         return orcid;
     }
@@ -341,15 +365,7 @@ public class OrcidSolrDocumentLegacy {
     public void setPrimaryRecord(String primaryRecord) {
         this.primaryRecord = primaryRecord;
     }
-
-    public List<String> getResearchResourceItemNames() {
-        return researchResourceItemNames;
-    }
-
-    public void setResearchResourceItemNames(List<String> researchResourceItemNames) {
-        this.researchResourceItemNames = researchResourceItemNames;
-    }
-
+    
     public List<String> getResearchResourceProposalTitles() {
         return researchResourceProposalTitles;
     }
@@ -660,8 +676,72 @@ public class OrcidSolrDocumentLegacy {
 
     public void setOrganisationNames(Map<String, Set<String>> organisationNames) {
         this.organisationNames = organisationNames;
-    }    
-    
+    }
+
+    public List<String> getExternalIdTypeAndValue() {
+        return externalIdTypeAndValue;
+    }
+
+    public void setExternalIdTypeAndValue(List<String> externalIdTypeAndValue) {
+        this.externalIdTypeAndValue = externalIdTypeAndValue;
+    }
+
+    public List<String> getExternalIdSourceReferences() {
+        return externalIdSourceReferences;
+    }
+
+    public void setExternalIdSourceReferences(List<String> externalIdSourceReferences) {
+        this.externalIdSourceReferences = externalIdSourceReferences;
+    }
+
+    public List<String> getResearchResourceItemNames() {
+        return researchResourceItemNames;
+    }
+
+    public void setResearchResourceItemNames(List<String> researchResourceItemNames) {
+        this.researchResourceItemNames = researchResourceItemNames;
+    }
+
+    public Set<String> getPeerReviewGroupId() {
+        return peerReviewGroupId;
+    }
+
+    public void setPeerReviewGroupId(Set<String> peerReviewGroupId) {
+        this.peerReviewGroupId = peerReviewGroupId;
+    }
+
+    public Set<String> getPeerReviewRole() {
+        return peerReviewRole;
+    }
+
+    public void setPeerReviewRole(Set<String> peerReviewRole) {
+        this.peerReviewRole = peerReviewRole;
+    }
+
+    public Set<String> getPeerReviewType() {
+        return peerReviewType;
+    }
+
+    public void setPeerReviewType(Set<String> peerReviewType) {
+        this.peerReviewType = peerReviewType;
+    }
+
+    public List<String> getPastInstitutionAffiliationNames() {
+        return pastInstitutionAffiliationNames;
+    }
+
+    public void setPastInstitutionAffiliationNames(List<String> pastInstitutionAffiliationNames) {
+        this.pastInstitutionAffiliationNames = pastInstitutionAffiliationNames;
+    }
+
+    public List<String> getCurrentInstitutionAffiliationNames() {
+        return currentInstitutionAffiliationNames;
+    }
+
+    public void setCurrentInstitutionAffiliationNames(List<String> currentInstitutionAffiliationNames) {
+        this.currentInstitutionAffiliationNames = currentInstitutionAffiliationNames;
+    }
+
     @Override
     public String toString() {
         return ToStringBuilder.reflectionToString(this);
@@ -670,6 +750,14 @@ public class OrcidSolrDocumentLegacy {
     private void generateCombinedGivenAndFamilyNames() {
         String givenAndFamilyNames = (givenNames == null) ? ((familyName == null) ? null : familyName) : (givenNames + ((familyName == null) ? "" : " " + familyName));
         setGivenAndFamilyNames(givenAndFamilyNames);
+    }
+
+    public String getBiography() {
+        return biography;
+    }
+
+    public void setBiography(String biography) {
+        this.biography = biography;
     }
 
     @Override
@@ -681,16 +769,20 @@ public class OrcidSolrDocumentLegacy {
         result = prime * result + ((asin == null) ? 0 : asin.hashCode());
         result = prime * result + ((asintld == null) ? 0 : asintld.hashCode());
         result = prime * result + ((bibcode == null) ? 0 : bibcode.hashCode());
+        result = prime * result + ((biography == null) ? 0 : biography.hashCode());
         result = prime * result + ((cba == null) ? 0 : cba.hashCode());
         result = prime * result + ((cit == null) ? 0 : cit.hashCode());
         result = prime * result + ((creditName == null) ? 0 : creditName.hashCode());
         result = prime * result + ((ctx == null) ? 0 : ctx.hashCode());
+        result = prime * result + ((currentInstitutionAffiliationNames == null) ? 0 : currentInstitutionAffiliationNames.hashCode());
         result = prime * result + ((digitalObjectIds == null) ? 0 : digitalObjectIds.hashCode());
         result = prime * result + ((eid == null) ? 0 : eid.hashCode());
         result = prime * result + ((emailAddresses == null) ? 0 : emailAddresses.hashCode());
         result = prime * result + ((ethos == null) ? 0 : ethos.hashCode());
         result = prime * result + ((externalIdReferences == null) ? 0 : externalIdReferences.hashCode());
+        result = prime * result + ((externalIdSourceReferences == null) ? 0 : externalIdSourceReferences.hashCode());
         result = prime * result + ((externalIdSources == null) ? 0 : externalIdSources.hashCode());
+        result = prime * result + ((externalIdTypeAndValue == null) ? 0 : externalIdTypeAndValue.hashCode());
         result = prime * result + ((familyName == null) ? 0 : familyName.hashCode());
         result = prime * result + ((fundingTitles == null) ? 0 : fundingTitles.hashCode());
         result = prime * result + ((givenAndFamilyNames == null) ? 0 : givenAndFamilyNames.hashCode());
@@ -714,15 +806,19 @@ public class OrcidSolrDocumentLegacy {
         result = prime * result + ((otherIdentifierType == null) ? 0 : otherIdentifierType.hashCode());
         result = prime * result + ((otherNames == null) ? 0 : otherNames.hashCode());
         result = prime * result + ((partOfIds == null) ? 0 : partOfIds.hashCode());
+        result = prime * result + ((pastInstitutionAffiliationNames == null) ? 0 : pastInstitutionAffiliationNames.hashCode());
         result = prime * result + ((pat == null) ? 0 : pat.hashCode());
+        result = prime * result + ((peerReviewGroupId == null) ? 0 : peerReviewGroupId.hashCode());
+        result = prime * result + ((peerReviewRole == null) ? 0 : peerReviewRole.hashCode());
+        result = prime * result + ((peerReviewType == null) ? 0 : peerReviewType.hashCode());
         result = prime * result + ((pmc == null) ? 0 : pmc.hashCode());
         result = prime * result + ((pmid == null) ? 0 : pmid.hashCode());
         result = prime * result + ((primaryRecord == null) ? 0 : primaryRecord.hashCode());
         result = prime * result + ((profileLastModifiedDate == null) ? 0 : profileLastModifiedDate.hashCode());
         result = prime * result + ((profileSubmissionDate == null) ? 0 : profileSubmissionDate.hashCode());
         result = prime * result + ((publicProfileMessage == null) ? 0 : publicProfileMessage.hashCode());
-        result = prime * result + ((researchResourceProposalTitles == null) ? 0 : researchResourceProposalTitles.hashCode());
         result = prime * result + ((researchResourceItemNames == null) ? 0 : researchResourceItemNames.hashCode());
+        result = prime * result + ((researchResourceProposalTitles == null) ? 0 : researchResourceProposalTitles.hashCode());
         result = prime * result + ((rfc == null) ? 0 : rfc.hashCode());
         result = prime * result + ((selfIds == null) ? 0 : selfIds.hashCode());
         result = prime * result + ((sourceWorkId == null) ? 0 : sourceWorkId.hashCode());
@@ -744,7 +840,7 @@ public class OrcidSolrDocumentLegacy {
             return false;
         if (getClass() != obj.getClass())
             return false;
-        OrcidSolrDocumentLegacy other = (OrcidSolrDocumentLegacy) obj;
+        OrcidSolrDocument other = (OrcidSolrDocument) obj;
         if (agr == null) {
             if (other.agr != null)
                 return false;
@@ -770,6 +866,11 @@ public class OrcidSolrDocumentLegacy {
                 return false;
         } else if (!bibcode.equals(other.bibcode))
             return false;
+        if (biography == null) {
+            if (other.biography != null)
+                return false;
+        } else if (!biography.equals(other.biography))
+            return false;
         if (cba == null) {
             if (other.cba != null)
                 return false;
@@ -789,6 +890,11 @@ public class OrcidSolrDocumentLegacy {
             if (other.ctx != null)
                 return false;
         } else if (!ctx.equals(other.ctx))
+            return false;
+        if (currentInstitutionAffiliationNames == null) {
+            if (other.currentInstitutionAffiliationNames != null)
+                return false;
+        } else if (!currentInstitutionAffiliationNames.equals(other.currentInstitutionAffiliationNames))
             return false;
         if (digitalObjectIds == null) {
             if (other.digitalObjectIds != null)
@@ -815,10 +921,20 @@ public class OrcidSolrDocumentLegacy {
                 return false;
         } else if (!externalIdReferences.equals(other.externalIdReferences))
             return false;
+        if (externalIdSourceReferences == null) {
+            if (other.externalIdSourceReferences != null)
+                return false;
+        } else if (!externalIdSourceReferences.equals(other.externalIdSourceReferences))
+            return false;
         if (externalIdSources == null) {
             if (other.externalIdSources != null)
                 return false;
         } else if (!externalIdSources.equals(other.externalIdSources))
+            return false;
+        if (externalIdTypeAndValue == null) {
+            if (other.externalIdTypeAndValue != null)
+                return false;
+        } else if (!externalIdTypeAndValue.equals(other.externalIdTypeAndValue))
             return false;
         if (familyName == null) {
             if (other.familyName != null)
@@ -935,10 +1051,30 @@ public class OrcidSolrDocumentLegacy {
                 return false;
         } else if (!partOfIds.equals(other.partOfIds))
             return false;
+        if (pastInstitutionAffiliationNames == null) {
+            if (other.pastInstitutionAffiliationNames != null)
+                return false;
+        } else if (!pastInstitutionAffiliationNames.equals(other.pastInstitutionAffiliationNames))
+            return false;
         if (pat == null) {
             if (other.pat != null)
                 return false;
         } else if (!pat.equals(other.pat))
+            return false;
+        if (peerReviewGroupId == null) {
+            if (other.peerReviewGroupId != null)
+                return false;
+        } else if (!peerReviewGroupId.equals(other.peerReviewGroupId))
+            return false;
+        if (peerReviewRole == null) {
+            if (other.peerReviewRole != null)
+                return false;
+        } else if (!peerReviewRole.equals(other.peerReviewRole))
+            return false;
+        if (peerReviewType == null) {
+            if (other.peerReviewType != null)
+                return false;
+        } else if (!peerReviewType.equals(other.peerReviewType))
             return false;
         if (pmc == null) {
             if (other.pmc != null)
@@ -970,15 +1106,15 @@ public class OrcidSolrDocumentLegacy {
                 return false;
         } else if (!publicProfileMessage.equals(other.publicProfileMessage))
             return false;
-        if (researchResourceProposalTitles == null) {
-            if (other.researchResourceProposalTitles != null)
-                return false;
-        } else if (!researchResourceProposalTitles.equals(other.researchResourceProposalTitles))
-            return false;
         if (researchResourceItemNames == null) {
             if (other.researchResourceItemNames != null)
                 return false;
         } else if (!researchResourceItemNames.equals(other.researchResourceItemNames))
+            return false;
+        if (researchResourceProposalTitles == null) {
+            if (other.researchResourceProposalTitles != null)
+                return false;
+        } else if (!researchResourceProposalTitles.equals(other.researchResourceProposalTitles))
             return false;
         if (rfc == null) {
             if (other.rfc != null)
@@ -1030,6 +1166,7 @@ public class OrcidSolrDocumentLegacy {
                 return false;
         } else if (!zbl.equals(other.zbl))
             return false;
+        
         return true;
     }
 }
