+++ orcid-core/src/test/java/org/orcid/core/solr/OrcidSolrOrgsClientTest.java
@@ -21,9 +21,9 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
+import org.orcid.core.utils.solr.entities.OrgDisambiguatedSolrDocument;
 import org.orcid.test.OrcidJUnit4ClassRunner;
 import org.orcid.test.TargetProxyHelper;
-import org.orcid.utils.solr.entities.OrgDisambiguatedSolrDocument;
 import org.springframework.test.context.ContextConfiguration;
 
 @RunWith(OrcidJUnit4ClassRunner.class)
